Estado Actual del Proyecto

+++ Assets/Scripts/Weapon Scripts/WeaponStats.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class WeaponStats : ScriptableObject
 {
     [Header("Weapon Stats")]
-    [SerializeField] private float _Damage;
+    [SerializeField] private int _Damage;
     [SerializeField] private float _Range;
     [SerializeField] private float _FireRate;
     [SerializeField] private float _ImpactForce;
@@ -14,7 +14,7 @@ public class WeaponStats : ScriptableObject
     [SerializeField] private float _ReloadTime;
     [SerializeField] private ParticleSystem _MuzzleFlash;
 
-    public float Damage { get => _Damage;}
+    public int Damage { get => _Damage;}
     public float Range { get => _Range;}
     public float FireRate { get => _FireRate;}
     public float ImpactForce { get => _ImpactForce;}
